+++ src/common/components/front-rich-components/calendar/index.ts
@@ -1,0 +1,1 @@
+export * from './calendar';
