+++ src/common/components/front-rich-components/index.ts
@@ -6,4 +6,4 @@ export * from './map-chart';
 export * from './video-player';
 export * from './bar-chart';
 export * from './line-chart';
-export * from './calendar';
+export * from './calendar/calendar';
